backend/requirements.txt
Django-3.2.18
asgiref-3.6.0
pytz-2023.3
sqlparse-0.4.4
typing-extensions-4.5.0
djangorestframework-3.14.0
Pillow-9.5.0